+++ packages/launcher/src/builtin/realm/LevelXTeamRoom.ts
@@ -44,7 +44,6 @@ export default (logger: AnyFunction, battle: (name: string) => ILevelBattle) => 
         constructor(public option: LevelOption) {}
 
         async update() {
-            this.logger(`${this.meta.name}: 更新关卡信息...`);
             const bits = await socket.bitSet(1000585, 2000036);
             const values = await socket.multiValue(12769, 12774, 20133);
             const pInfos = await socket.playerInfo(1197);
@@ -58,27 +57,22 @@ export default (logger: AnyFunction, battle: (name: string) => ILevelBattle) => 
             this.data.weeklyCompletedCount = values[2];
 
             if (this.data.weeklyRewardReceived) {
-                this.logger(`${this.meta.name}: 日任完成`);
                 return LevelAction.STOP;
             }
 
             if (!this.data.weeklyRewardReceived && this.data.weeklyCompletedCount >= 5) {
-                this.logger(`${this.meta.name}: 领取每周奖励`);
                 return 'award_weekly';
             }
 
             if (this.data.dailyRewardReceived) {
-                this.logger(`${this.meta.name}: 日任完成`);
                 return LevelAction.STOP;
             }
 
             if (!this.data.dailyRewardReceived && this.data.dailyMinRound > 0) {
-                this.logger(`${this.meta.name}: 领取每日奖励`);
                 return LevelAction.AWARD;
             }
 
             if (this.data.dailyMinRound === 0 && (this.data.remainingTimes > 0 || this.data.open)) {
-                this.logger(`${this.meta.name}: 进入战斗`);
                 if (this.data.open) {
                     return LevelAction.BATTLE;
                 } else {